buildscripts: make the build more parallel and less racy

+++ buildscripts/build.sh
@@ -181,27 +181,29 @@ find ./toolchain/$ARCH/ -iname "libc++_shared.so" -exec cp "{}" ../app/src/main/
 
 $NDK_TRIPLET-strip ../app/src/main/jniLibs/$ABI/*.so
 
-echo "==> Deploying resources"
+if [[ $ARCH = "arm" ]]; then
+	echo "==> Deploying resources"
 
-DST=$DIR/../app/src/main/assets/libopenmw/
-SRC=build/$ARCH/openmw-prefix/src/openmw-build/
+	DST=$DIR/../app/src/main/assets/libopenmw/
+	SRC=build/$ARCH/openmw-prefix/src/openmw-build/
 
-rm -rf "$DST" && mkdir -p "$DST"
+	rm -rf "$DST" && mkdir -p "$DST"
 
-# resources
-cp -r "$SRC/resources" "$DST"
+	# resources
+	cp -r "$SRC/resources" "$DST"
 
-# global config
-mkdir -p "$DST/openmw/"
-cp "$SRC/settings-default.cfg" "$DST/openmw/"
-cp "$SRC/gamecontrollerdb.txt" "$DST/openmw/"
+	# global config
+	mkdir -p "$DST/openmw/"
+	cp "$SRC/settings-default.cfg" "$DST/openmw/"
+	cp "$SRC/gamecontrollerdb.txt" "$DST/openmw/"
 
-# local config
-mkdir -p "$DST/config/openmw/"
-# TODO: do we really need this twice?
-cp "$SRC/gamecontrollerdb.txt" "$DST/config/openmw/"
-cp "$DIR/../app/openmw-base.cfg" "$DST/config/openmw/openmw.cfg"
-cp "$DIR/../app/settings-base.cfg" "$DST/config/openmw/settings.cfg"
+	# local config
+	mkdir -p "$DST/config/openmw/"
+	# TODO: do we really need this twice?
+	cp "$SRC/gamecontrollerdb.txt" "$DST/config/openmw/"
+	cp "$DIR/../app/openmw-base.cfg" "$DST/config/openmw/openmw.cfg"
+	cp "$DIR/../app/settings-base.cfg" "$DST/config/openmw/settings.cfg"
+fi
 
 echo "==> Making your debugging life easier"
